refactor: :art: Move getPokemons function to helpers folder
